requirements.txt
flask==3.0.0
python-dateutil==2.8.2
Werkzeug==3.0.1